+++ Forms/PreviewForm.vb
@@ -1024,7 +1024,7 @@ Public Class PreviewForm
     Protected Overrides Sub OnLoad(args As EventArgs)
         MyBase.OnLoad(args)
 
-        PreviewScript.Synchronize(True, True, True, TextEncoding.EncodingOfProcess)
+        PreviewScript.Synchronize(True, True, True)
         FrameServer = FrameServerFactory.Create(PreviewScript.Path)
         Renderer = New VideoRenderer(pnVideo, FrameServer)
         Renderer.Info = PreviewScript.OriginalInfo