+++ General/Audio.vb
@@ -161,7 +161,7 @@ Public Class Audio
         Dim d As New VideoScript
         d.Filters.AddRange(p.Script.Filters)
         Dim wavPath = p.TempDir + ap.File.Base + "_cut_na.wav"
-        d.Path = (p.TempDir + ap.File.Base + "_cut_na.avs").LongPathPrefix
+        d.Path = p.TempDir + ap.File.Base + "_cut_na.avs"
         d.Filters.Insert(1, New VideoFilter(GetNicAudioCode(ap)))
 
         If ap.Channels = 2 Then
@@ -208,7 +208,7 @@ Public Class Audio
         d.Filters.AddRange(p.Script.Filters)
         d.RemoveFilter("Cutting")
         Dim outPath = p.TempDir + ap.File.Base + "_DecodeNicAudio." + ap.ConvertExt
-        d.Path = (p.TempDir + ap.File.Base + "_DecodeNicAudio.avs").LongPathPrefix
+        d.Path = p.TempDir + ap.File.Base + "_DecodeNicAudio.avs"
         d.Filters.Insert(1, New VideoFilter(GetNicAudioCode(ap)))
 
         If ap.Channels = 2 Then
@@ -342,13 +342,13 @@ Public Class Audio
             base += ap.GetTrackID
         End If
 
-        Dim outPath = (p.TempDir + base + "." + ap.ConvertExt).LongPathPrefix
+        Dim outPath = p.TempDir + base + "." + ap.ConvertExt
 
         If ap.File = outPath Then
             outPath += "." + ap.ConvertExt
         End If
 
-        Dim args = "-i " + ap.File.LongPathPrefix.Escape
+        Dim args = "-i " + ap.File.Escape
 
         If Not ap.Stream Is Nothing Then
             args += " -map 0:" & ap.Stream.StreamOrder
@@ -443,7 +443,7 @@ Public Class Audio
         d.Filters.AddRange(p.Script.Filters)
         d.RemoveFilter("Cutting")
         Dim outPath = p.TempDir + ap.File.Base + "_convDSS." + ap.ConvertExt
-        d.Path = (p.TempDir + ap.File.Base + "_DecDSS.avs").LongPathPrefix
+        d.Path = p.TempDir + ap.File.Base + "_DecDSS.avs"
         d.Filters.Insert(1, New VideoFilter("AudioDub(last,DirectShowSource(""" + ap.File + """, video=false))"))
 
         If ap.Channels = 2 Then
@@ -488,7 +488,7 @@ Public Class Audio
         d.Filters.AddRange(p.Script.Filters)
         d.RemoveFilter("Cutting")
         Dim outPath = p.TempDir + ap.File.Base + "_convFFAudioSource." + ap.ConvertExt
-        d.Path = (p.TempDir + ap.File.Base + "_DecodeFFAudioSource.avs").LongPathPrefix
+        d.Path = p.TempDir + ap.File.Base + "_DecodeFFAudioSource.avs"
         d.Filters.Insert(1, New VideoFilter("AudioDub(last,FFAudioSource(""" + ap.File + """, cachefile=""" + cachefile + """))"))
 
         If ap.Channels = 2 Then
@@ -530,7 +530,7 @@ Public Class Audio
         Dim d As New VideoScript
         d.Filters.AddRange(p.Script.Filters)
         Dim wavPath = p.TempDir + ap.File.Base + "_cut_ds.wav"
-        d.Path = (p.TempDir + ap.File.Base + "_cut_ds.avs").LongPathPrefix
+        d.Path = p.TempDir + ap.File.Base + "_cut_ds.avs"
         d.Filters.Insert(1, New VideoFilter("AudioDub(last,DirectShowSource(""" + ap.File + """, video=false))"))
 
         If ap.Channels = 2 Then
@@ -574,7 +574,7 @@ Public Class Audio
         Dim d As New VideoScript
         d.Filters.AddRange(p.Script.Filters)
         Dim wavPath = p.TempDir + ap.File.Base + "_cut_ff.wav"
-        d.Path = (p.TempDir + ap.File.Base + "_cut_ff.avs").LongPathPrefix
+        d.Path = p.TempDir + ap.File.Base + "_cut_ff.avs"
         d.Filters.Insert(1, New VideoFilter("AudioDub(last,FFAudioSource(""" + ap.File + """, cachefile=""" + cachefile + """))"))
 
         If ap.Channels = 2 Then